Bullets destroyed when colliding with walls and player and enemy

+++ Unity Game 1/Assets/Scripts/Spawner.cs
@@ -6,7 +6,7 @@ public class Spawner : MonoBehaviour {
 
     public bool Active;
 
-    public int SpawnSpeed;
+    public float SpawnSpeed;
     public GameObject EnemyPrefab;
     private GameObject enemy;
     private float timerCounter;
